+++ tfsnippet/scaffold/__init__.py
@@ -1,5 +1,4 @@
 from .checkpoint import *
-from .early_stopping_ import *
 from .event_keys import *
 from .logging_ import *
 from .scheduled_var import *
@@ -7,7 +6,6 @@ from .train_loop_ import *
 
 __all__ = [
     'AnnealingVariable', 'CheckpointSavableObject', 'CheckpointSaver',
-    'DefaultMetricFormatter', 'EarlyStopping', 'EarlyStoppingContext',
-    'EventKeys', 'MetricFormatter', 'MetricLogger', 'ScheduledVariable',
-    'TrainLoop', 'early_stopping', 'summarize_variables',
+    'DefaultMetricFormatter', 'EventKeys', 'MetricFormatter', 'MetricLogger',
+    'ScheduledVariable', 'TrainLoop', 'summarize_variables',
 ]